refactor(process): move commonly used order and color methods to separate files to clearify shared usage and to avoid duplication and enforce consistency
